testing CoRs computation
change float to double precision for Skeleton::simil 
remove unused vertex.h
compute a small subset of the CoRs at runtime for testing purposes

+++ skeleton.h
@@ -31,7 +31,7 @@ public:
     Skeleton();
     ~Skeleton();
     inline float* weightsAt(uint vertex) { return weights + (vertex*edges.size());}
-    float simil(uint vertexInd, Triangle t);
+    double simil(uint vertexInd, Triangle t);
     bool parseSkelFile(const std::string& file);
     std::vector<QVector3D> getSkelLines();
     uint getNumberBones() {return edges.size();}
